LayoutTests/http/tests/permissions/storage-access-permissions-query-ephemeral-expected.txt

This test checks that navigator.permissions.query() works for storage-access permission in an ephemeral session.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS state is "prompt"
PASS state is "granted"
PASS state is "prompt"
PASS successfullyParsed is true

TEST COMPLETE
